refactor: rename sign-up routes to authentication routes
