+++ src/main/adapters/index.ts
@@ -1,1 +1,1 @@
-export * from './express-route'
+export * from './adapt-route'
