+++ src/main/factories/validations/login/index.ts
@@ -5,7 +5,7 @@ import {
   ValidationComposite,
 } from '@/presentation/helpers/validators'
 
-import { EmailValidatorAdapter } from '@/utils/email-validator/validator'
+import { EmailValidatorAdapter } from '@/main/adapters/email-validator'
 
 export const makeLoginValidation = (): IValidation => {
   const validations: IValidation[] = []
